Tuned percentEscapedToRealBytes() function
The percentEscapedToRealBytes() function in the Http namespace did not
properly convert '+' characters into spaces as the standard dictates. I
also changed the return type to size_t and eliminated a duplicate line
of code.

+++ include/fastcgi++/http.hpp
@@ -480,7 +480,7 @@ namespace Fastcgipp
 		 * @param[out] destination Pointer to the section of memory to write the converted string to
 		 * @return Actual size of the new string
 		 */
-		template <class charT> int percentEscapedToRealBytes(const charT* source, charT* destination, size_t size);
+		template <class charT> size_t percentEscapedToRealBytes(const charT* source, charT* destination, size_t size);
 
 		/** 
 		 * @brief List of characters in order for Base64 encoding.
